requirements.txt
pandas
inflection
streamlit
folium
streamlit_folium
plotly.express